analyPage/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.6.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==3.0.5
idna==2.9
joblib==0.14.1
JPype1==0.7.3
konlpy==0.5.2
lxml==4.5.0
numpy==1.18.3
oauthlib==3.1.0
pandas==1.0.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tweepy==3.8.0
urllib3==1.25.9
